requirements.txt
opencv-python
python-chess
mss
numpy
pyautogui
Pillow
requests